requirements.txt
streamlit
openai
llama-index==0.9.29
nltk
pypdf
llama-hub
unstructured==0.12.0
fastapi==0.109.0
uvicorn==0.27.0.post1
pymongo==4.6.1
python-multipart==0.0.6
python-dotenv==1.0.1
motor==3.3.2
beanie==1.25.0
fastapi-jwt-auth==0.5.0
pydantic-settings==2.1.0
passlib==1.7.4
